Fix Vector coordinate size

+++ src/sdk/Vector.cpp
@@ -5,28 +5,28 @@ uint8_t operator==(const Vector &v1, const Vector &v2)
 	return v1.x == v2.x && v1.y == v2.y;
 }
 
-Vector::Vector(uint16_t x, uint16_t y) : x(x),
-										 y(y)
+Vector::Vector(uint8_t x, uint8_t y) : x(x),
+                                       y(y)
 {
 
 }
 
-void Vector::SetX(uint16_t x)
+void Vector::SetX(uint8_t x)
 {
 	this->x = x;
 }
 
-void Vector::SetY(uint16_t y)
+void Vector::SetY(uint8_t y)
 {
 	this->y = y;
 }
 
-uint16_t Vector::GetX() const
+uint8_t Vector::GetX() const
 {
 	return x;
 }
 
-uint16_t Vector::GetY() const
+uint8_t Vector::GetY() const
 {
 	return y;
 }
